Move hardware detection to AutomationCore
